requirements.txt

Django==2.2.5
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.10.3
mongoengine==0.18.2